changed TimeProvider and Transitions to use a float with seconds instead an int/long with millis

+++ animation4j-componentsengine/src/main/java/com/gemserk/animation4j/componentsengine/properties/DelayedProperty.java
@@ -14,7 +14,7 @@ public class DelayedProperty<T> implements Property<T> {
 
 	private T delayedValue;
 
-	private long timeWhenSet = 0;
+	private float timeWhenSet = 0;
 
 	private T value;
